+++ docs/developer/plugin-list.asciidoc
@@ -134,7 +134,7 @@ This API doesn't support angular, for registering angular dev tools, bootstrap a
 |Expression Image plugin adds an image renderer to the expression plugin. The renderer will display the given image.
 
 
-|{kib-repo}blob/{branch}/src/plugins/chart_expressions/expression_legacy_metric/README.md[expressionLegacyMetricVis]
+|{kib-repo}blob/{branch}/src/platform/plugins/shared/chart_expressions/expression_legacy_metric/README.md[expressionLegacyMetricVis]
 |Expression MetricVis plugin adds a metric renderer and function to the expression plugin. The renderer will display the metric chart.
