+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/.i18nrc.json
