+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/__mocks__/format_service.ts
