+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/__mocks__/palette_service.ts
