+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/__stories__/metric_renderer.stories.tsx
