+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/format_service.ts
