+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/services/palette_service.ts
