+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/public/utils/index.ts
