+++ src/platform/plugins/shared/chart_expressions/expression_legacy_metric/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../../tsconfig.base.json",
+  "extends": "../../../../../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "target/types",
     "isolatedModules": true
