+++ tsconfig.base.json
@@ -908,8 +908,8 @@
       "@kbn/expression-heatmap-plugin/*": ["src/platform/plugins/shared/chart_expressions/expression_heatmap/*"],
       "@kbn/expression-image-plugin": ["src/platform/plugins/shared/expression_image"],
       "@kbn/expression-image-plugin/*": ["src/platform/plugins/shared/expression_image/*"],
-      "@kbn/expression-legacy-metric-vis-plugin": ["src/plugins/chart_expressions/expression_legacy_metric"],
-      "@kbn/expression-legacy-metric-vis-plugin/*": ["src/plugins/chart_expressions/expression_legacy_metric/*"],
+      "@kbn/expression-legacy-metric-vis-plugin": ["src/platform/plugins/shared/chart_expressions/expression_legacy_metric"],
+      "@kbn/expression-legacy-metric-vis-plugin/*": ["src/platform/plugins/shared/chart_expressions/expression_legacy_metric/*"],
       "@kbn/expression-metric-plugin": ["src/platform/plugins/shared/expression_metric"],
       "@kbn/expression-metric-plugin/*": ["src/platform/plugins/shared/expression_metric/*"],
       "@kbn/expression-metric-vis-plugin": ["src/plugins/chart_expressions/expression_metric"],
